+++ components/decimal/src/preferences.rs
@@ -60,22 +60,3 @@ prefs_convert!(
     CompactDecimalFormatterPreferences,
     icu_plurals::PluralRulesPreferences
 );
-
-use icu_provider::prelude::*;
-
-impl DecimalFormatterPreferences {
-    pub(crate) fn nu_id<'a>(
-        &'a self,
-        locale: &'a DataLocale,
-    ) -> Option<DataIdentifierBorrowed<'a>> {
-        self.numbering_system
-            .as_ref()
-            .map(|s| s.as_str())
-            .map(|nu| {
-                DataIdentifierBorrowed::for_marker_attributes_and_locale(
-                    DataMarkerAttributes::from_str_or_panic(nu),
-                    locale,
-                )
-            })
-    }
-}
